[#447] 해피지수 프로그래스바에 반영

+++ Mohaeng/Mohaeng/Sources/ViewControllers/PopUp/HomeHappy/HappyPopUpViewController.swift
@@ -40,11 +40,11 @@ class HappyPopUpViewController: UIViewController {
     func initData() {
         levelLabel.text = "Lv.\(data.level)"
         happyPercentLabel.text = "해피지수 \(data.happy)/\(data.fullHappy) (\(getHappyPercent())%)"
-        progressView.setProgress(Float(data.happy / data.fullHappy), animated: false)
+        progressView.setProgress(Float(data.happy) / Float(data.fullHappy), animated: false)
     }
     
     func getHappyPercent() -> Int {
-        return data.happy / data.fullHappy * 100
+        return Int(Float(data.happy) / Float(data.fullHappy) * 100)
     }
 
     // MARK: - @IBAction Functions
